avoid cast error, standardize types and calls

+++ src/students/students.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Student } from './student.model';
@@ -9,19 +13,19 @@ export class StudentsService {
     @InjectModel('Student') private readonly studentModel: Model<Student>,
   ) {}
 
-  async createStudent(student: Student) {
+  async createStudent(student: Student): Promise<{ id: string }> {
     const studentModel = new this.studentModel({
       name: student.name,
       tia: student.tia,
       course: student.course,
     });
     const result = await studentModel.save();
-    return result.id as string;
+    return { id: result.id };
   }
 
-  async readAll() {
+  async readAll(): Promise<Student[]> {
     const students = await this.studentModel.find().exec();
-    return students.map(student => ({
+    return students.map((student) => ({
       id: student.id,
       name: student.name,
       tia: student.tia,
@@ -30,6 +34,10 @@ export class StudentsService {
   }
 
   async getStudentByTIA(tia: number): Promise<Student> {
+    if (isNaN(Number(tia))) {
+      throw new BadRequestException('TIA must be a number');
+    }
+
     const student = await this.studentModel.findOne({ tia: tia });
     if (!student) {
       throw new NotFoundException('Could not find the student.');
@@ -42,11 +50,15 @@ export class StudentsService {
     };
   }
 
-  async deleteStudent(tia: number) {
+  async deleteStudent(tia: number): Promise<{ message: string }> {
+    if (isNaN(Number(tia))) {
+      throw new BadRequestException('TIA must be a number');
+    }
     const result = await this.studentModel.deleteOne({ tia: tia }).exec();
     if (result.deletedCount === 0) {
       throw new NotFoundException('Could not remove student.');
     }
+    return { message: `Student with tia ${tia} deleted` };
   }
 
   async updateStudent(student: Student): Promise<Student> {
